+++ src/parser/lang.rs
@@ -7,7 +7,7 @@ pub fn symbol(text: &str) -> Result<Tree, ParserError> {
     let poses = Location::new_line_poses(text);
 
     Ok(Tree::from(state.into_bfs().map(|x| NodeData {
-        location: Location::from_position(x.position, x.symbol.len(), &poses),
+        location: Location::from_position(text, x.position, x.symbol.len(), &poses),
         symbol:   x.symbol,
     })))
 }
@@ -20,7 +20,7 @@ pub fn terms(text: &str) -> Result<Vec<Tree>, ParserError> {
         .into_iter()
         .map(|s| {
             Tree::from(s.into_bfs().map(|x| NodeData {
-                location: Location::from_position(x.position, x.symbol.len(), &poses),
+                location: Location::from_position(text, x.position, x.symbol.len(), &poses),
                 symbol:   x.symbol,
             }))
         })
@@ -32,7 +32,7 @@ pub fn lang_rule(text: &str) -> Result<Tree, ParserError> {
     let poses = Location::new_line_poses(text);
 
     Ok(Tree::from(state.into_bfs().map(|x| NodeData {
-        location: Location::from_position(x.position, x.symbol.len(), &poses),
+        location: Location::from_position(text, x.position, x.symbol.len(), &poses),
         symbol:   x.symbol,
     })))
 }
@@ -42,7 +42,7 @@ pub fn task(text: &str) -> Result<Tree, ParserError> {
     let poses = Location::new_line_poses(text);
 
     Ok(Tree::from(state.into_bfs().map(|x| NodeData {
-        location: Location::from_position(x.position, x.symbol.len(), &poses),
+        location: Location::from_position(text, x.position, x.symbol.len(), &poses),
         symbol:   x.symbol,
     })))
 }
@@ -55,7 +55,7 @@ pub fn any(text: &str) -> Result<Vec<Tree>, ParserError> {
         .into_iter()
         .map(|s| {
             Tree::from(s.into_bfs().map(|x| NodeData {
-                location: Location::from_position(x.position, x.symbol.len(), &poses),
+                location: Location::from_position(text, x.position, x.symbol.len(), &poses),
                 symbol:   x.symbol,
             }))
         })